agents/requirements.txt
langchain==0.3.19
langchain-core==0.3.35
langchain-google-genai
python-dotenv
requests==2.31.0
yfinance==0.2.40
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0